Separate statistics code to each chain type directory
